Remove unnecessary module check on redeclaredInMultipleModulesFunc1
For the pursposes of code-completion, we don't really care what module
it came from, and it is likely that some underlying non-determinism is
causing the recent rare failures of this test. We should fix the
non-determinism wherever it is, but for now we can attempt to avoid the
test failure.

rdar://problem/28495403