Simpler implementation for multiple matching patterns in case with variable bindings
Backed off of the implementation with a shared body block, which got
messy really fast with multiple patterns containing address-only values
to pass in, as it led to separated stack alloc/cleanup requirements and
passing stack alloc’d values as block args, aliasing of addresses for
stack cleanup verifying, etc. A mess.

This now does the conservative thing of repeating the body in each
matching pattern block.