SILCombine: handle upcasts when converting from alloc_ref_dynamic to alloc_ref
We transform an "alloc_ref_dynamic(metatype @thick $T)" to an "alloc_ref $T"
But in some cases there can be an upcast in-between.
Now we also do "alloc_ref_dynamic(upcast(metatype @thick $Derived) to  $Base)" -> "upcast(alloc_ref $Derived) to $Base"

This peephole optimization is needed for the upcoming change of using tail allocated arrays in ManagedBuffer.