[CMake] Check target, not host, for Info.plist
The CMake variable `APPLE` is set to true when building on a macOS
host, which means that logic to include linker flags for application
extensions, etc., will be enabled on a macOS host. That is, they will
be enabled even when cross-compiling the standard library to a
non-Apple platform, such as Android.

Instead of using `APPLE` to check the host OS, check whether the target
is an Apple platform.