Revert "[CMake] Fix dependency on codesign steps"
This reverts commit 10506e00c09d97342ecce687ef8ef6aa43ebda1d.

There are dylibs in lib/swift/macosx/x86_64/ and lib/swift/macosx/, and
with this patch the dylibs in the latter path don't get updated when the
code changes and are stale.

I'm not sure exactly how to correct this patch so let's just try again
when it updates both sets of dylibs.

rdar://28660201