[code-completion] Handle generic function refs converted to non-generic function type
When checking a completion of generic function type against a context
with function type, we want to allow the conversion if we can match up
the constraints.  Add a new entry point to do the conversion check after
opening any generic types.

Note: for now this only works in one direction: assigning a generic
function to a non-generic context.  The problem with the other direction
is that we currently filter out all generic types when computing the
ExpectedType(s) for the context.