[CMake] Replace APPLE with CMAKE_HOST_APPLE
In CMake `APPLE` is defined if you are *targeting* an Apple platform (or
at least if CMake thinks you are). If your host is an Apple platform
CMake defines `CMAKE_HOST_APPLE`.

Swift's build system is rolling its own cross-compiling support, so checks such
as `APPLE` do not function properly, and instead are true when the
*host* is `APPLE`. To make that fact explicit, change these variables to
`CMAKE_HOST_APPLE`.