(wip) IRGen: Include witness tables as part of generic metadata patterns' keys.
The conformance chosen is a significant part of the type, and can differ within a process if multiple modules independently evolve their own conformances of the same type to the same protocol. We used to unique by witness table, but that seems to have been accidentally lost due to other refactorings. The lack of uniquing covered up the problem in rdar://problem/28022201, because we would instantiate some types' metadata with a garbage witness table, then intermingle it with metadata with the correct witness table. Properly uniquing by conformance would make the type mismatch much more apparent at runtime.

(wip) because this exposes a crash in some tests because we fail to unique witness tables emitted for conformances implicitly generated by the Clang importer at runtime, leading to erroneously different Set and Dictionary instantiations on C types at runtime. We need to fix rdar://problem/18884333 before fixing this.