properly expose == to objc in test
This test was relying on an impl detail that was never guaranteed and
no longer holds (multiple bridgings of a dict had equal identity). Based
on how the code in the test is written, it appears to be an accident.
This change makes the test "make sense" without the impl detail.