Revert "[cmake] Add the option SWIFT_BUILD_RUNTIME_WITH_HOST_LIBLTO."
This reverts commit 23987b26f3c53d471034311098bd629971482183.

I originally thought that I was going to need this for LTO work. Now it is just
dead code.

Even if we were to enable this though it wouldn't fix the underlying issue since
swiftc would still emit bitcode that may be incompatible with the host compiler.

So reverting this makes the most sense.