[semantic-arc] When parsing/deserializing SILInstructions, we know that ResultVal is always a SILInstruction, so just make it a SILInstruction instead of a ValueBase.
As I was preparing a semantic arc commit that worked in this area, I realized I
needed ResultVal to be a SILInstruction, not a ValueBase. As I prepared to
perform a dyn_cast, I realized that ResulVal is always a SILInstruction
anyways... so why not just use the right type in the first place... Thus this
commit.

*sigh*

rdar://28685236