docs: description of new mangling scheme
These are the main changes:

1) Change the order of the mangling to a post-fix like structure.

This is the biggest change.
It will help to get more common prefixes in the mangled names to optimize the trie in the macho object files.
The length of the mangled names will mostly stay the same but the order of 'operands' inside the mangling is more or less reversed.
This change also required to use different 'operator' characters in some cases.

2) Word-substitutions

Similar to the S-substitutions, but finer grained. See section 'Identifiers'.
Reduces the size of mangled names in general.

3) Combined substitutions

A more efficient way to mangle multiple S-substitutions.
Reduces the size of mangled names with lots of substitutions, e.g. specialized functions.

4) Change the '_T' prefix to '_S'

(on John's request)
Because it's basically a new mangling scheme.