[sourcekit] Introduce a lazy representation for DocumentStructure responses
Previously, document structure responses were created using XPC
dictionaries, but this forced deserialization of the large nested
dictionaries no matter what the access pattern was.  The new format uses
our custom buffer machinery, and only deserializes dictionaries that are
actually needed.

The performance improvement is around 25% for an editor.open, and 35-40%
for an editor.replacetext request in the large files I've tested. The
performance improvements are in serialization (inside the service),
deserialization (client), and disposal of the response object (client).

rdar://problem/28789756